Moved attributes, ability and animation classes in correct folders
