+++ GameBaseFramework.uplugin
@@ -48,10 +48,6 @@
       "Enabled": true
     },
     {
-      "Name": "GASExtensions",
-      "Enabled": true
-    },
-    {
       "Name": "GameFeatures",
       "Enabled": true
     },
