+++ Source/GameBaseFramework/Private/Animation/GBFAnimLayerSelectionSet.cpp
@@ -1,4 +1,4 @@
-#include "Animations/GBFAnimLayerSelectionSet.h"
+#include "Animation/GBFAnimLayerSelectionSet.h"
 
 #include <Animation/AnimInstance.h>
