+++ Source/GameBaseFramework/Private/Characters/Components/GBFHealthComponent.cpp
@@ -1,6 +1,6 @@
 ﻿#include "Characters/Components/GBFHealthComponent.h"
 
-#include "GAS/GBFHealthAttributeSet.h"
+#include "GAS/Attributes/GBFHealthAttributeSet.h"
 #include "GAS/Components/GASExtAbilitySystemComponent.h"
 #include "GBFLog.h"
 #include "Log/CoreExtLog.h"
