+++ Source/GameBaseFramework/Private/GAS/Components/GASExtAbilitySystemComponent.cpp
@@ -1,9 +1,9 @@
 #include "GAS/Components/GASExtAbilitySystemComponent.h"
 
+#include "Animation/GASExtAnimInstance.h"
 #include "DVEDataValidator.h"
 #include "GAS/Abilities/GASExtAbilityTagRelationshipMapping.h"
 #include "GAS/Abilities/GASExtGameplayAbility.h"
-#include "GAS/Animation/GASExtAnimInstance.h"
 
 #include <AbilitySystemLog.h>
 #include <Animation/AnimInstance.h>
