+++ Source/GameBaseFramework/Public/Interaction/Abilities/GBFGameplayAbility_Interact.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "GAS/GBFGameplayAbility.h"
+#include "GAS/Abilities/GBFGameplayAbility.h"
 
 #include "GBFGameplayAbility_Interact.generated.h"
